+++ docs/src/routes.js
@@ -25,13 +25,6 @@ const routes = (
                 }}
                 path="colors"
             />
-            <Route
-                getComponent={(location, callback) => {
-                    import('./templates/page.js')
-                        .then(module => callback(null, module.default));
-                }}
-                path="page"
-            />
         </Route>
 
         <Route path="/elements">
@@ -289,6 +282,13 @@ const routes = (
                     }}
                     path="wing"
                 />
+                <Route
+                    getComponent={(location, callback) => {
+                        import('./modules/drawerDetails.js')
+                            .then(module => callback(null, module.default));
+                    }}
+                    path="details"
+                />
             </Route>
             <Route
                 getComponent={(location, callback) => {
@@ -297,6 +297,21 @@ const routes = (
                 }}
                 path="drawer-deprecated"
             />
+            <Route path="page">
+                <IndexRoute
+                    getComponent={(location, callback) => {
+                        import('./modules/page.js')
+                            .then(module => callback(null, module.default));
+                    }}
+                />
+                <Route
+                    getComponent={(location, callback) => {
+                        import('./modules/pageDemo.js')
+                            .then(module => callback(null, module.default));
+                    }}
+                    path="demo"
+                />
+            </Route>
             <Route
                 getComponent={(location, callback) => {
                     import('./modules/dropdown.js')