+++ docs/src/shared/configureReduxStore.js
@@ -1,0 +1,15 @@
+import { applyMiddleware, createStore } from 'redux';
+import rootReducer from '../reducers.js';
+import thunkMiddleware from 'redux-thunk';
+
+const middlewares = [ thunkMiddleware ];
+const enhancer = applyMiddleware(...middlewares);
+
+const configureReduxStore = () => {
+    return createStore(
+        rootReducer,
+        enhancer
+    );
+};
+
+export const appReduxStore = configureReduxStore();